requirements.txt
bidict==0.22.1
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
EasyProcess==1.1
entrypoint2==1.1
Flask==2.3.3
Flask-SocketIO==5.3.6
greenlet==2.0.2
h11==0.14.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
numpy==1.26.0
opencv-python==4.8.0.76
Pillow==10.0.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-engineio==4.7.1
python-socketio==5.9.0
pytweening==1.0.7
rubicon-objc==0.4.6
simple-websocket==0.10.1
six==1.16.0
Werkzeug==2.3.7
wsproto==1.2.0
zipp==3.17.0
